refactor(telemetry): Extract the export-independent aspects of prometheus.js
